backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
mysqlclient==2.2.4
python-dotenv==1.0.1
face-recognition==1.3.0
numpy==1.26.4